One step towards better component models genericity
Dispatch to the low-level function using the process type as first arg now instead of the parameter in the parametric component models type. Careful: also changed the parameter name for energy balance from energy to energy_balance to be more consistant with the function name. And gs function into stomatal_conductance function.